+++ mnelab/dialogs/historydialog.py
@@ -1,0 +1,18 @@
+from PyQt5.QtWidgets import (QDialog, QVBoxLayout, QPlainTextEdit,
+                             QDialogButtonBox)
+
+
+class HistoryDialog(QDialog):
+    def __init__(self, parent, history):
+        super().__init__(parent=parent)
+        self.setWindowTitle("History")
+        layout = QVBoxLayout()
+        text = QPlainTextEdit()
+        text.setReadOnly(True)
+        text.setPlainText(history)
+        layout.addWidget(text)
+        buttonbox = QDialogButtonBox(QDialogButtonBox.Ok)
+        layout.addWidget(buttonbox)
+        self.setLayout(layout)
+        buttonbox.accepted.connect(self.accept)
+        self.resize(700, 500)